+++ dl4cv/experimental/solver_idea2.py
