+++ dl4cv/experimental/train_idea1.py
@@ -4,7 +4,7 @@ import torch.nn as nn
 from dl4cv.dataset_utils import ForwardAndMirroredDatasetRAM
 
 from dl4cv.models.models import AutoEncoderPV
-from dl4cv.explicitDisentanglement.solver import Solver
+from dl4cv.experimental.solver import Solver
 from torch.utils.data import DataLoader, SequentialSampler
 
 config = {
