+++ dl4cv/experimental/train_idea2.py
@@ -4,7 +4,7 @@ import torch.nn as nn
 from dl4cv.dataset_utils import ForwardDatasetIdea2RAM
 
 from dl4cv.models.models import AutoEncoderIdea2
-from dl4cv.explicitDisentanglement.solver_idea2 import Solver
+from dl4cv.experimental.solver_idea2 import Solver
 from torch.utils.data import DataLoader, SequentialSampler
 
 config = {
